fix: make type of amount field float

+++ app/database/postgre/models/crypto_transaction.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import ForeignKey, String, Integer, Float, TIMESTAMP
+from sqlalchemy import ForeignKey, String, Float, TIMESTAMP
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.database.postgre import db, utcnow
@@ -11,7 +11,7 @@ class CryptoTransaction(db.Model):
 
     ID: Mapped[str] = mapped_column(String(36), primary_key=True)
     ticker: Mapped[str] = mapped_column(ForeignKey("CryptoCurrency.ticker"))
-    amount: Mapped[int] = mapped_column(Integer)
+    amount: Mapped[float] = mapped_column(Float)
     type_: Mapped[str] = mapped_column(String(8))
     time: Mapped[datetime] = mapped_column(TIMESTAMP, server_default=utcnow())
     price: Mapped[float] = mapped_column(Float)
@@ -22,7 +22,7 @@ class CryptoTransaction(db.Model):
         *,
         ID: str,
         ticker: str,
-        amount: int,
+        amount: float,
         type_: str,
         price: float,
         user_id: str
